pyproject.toml
[project]
name = "draiver_analysis"
version = "0.1.0"
description = "Data analysis and Google Sheets API client for Projeto Draiver"
authors = [
  { name = "Gabriel Canto" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "pandas>=2.2.3,<3.0",
  "gspread>=6.2.0,<7.0",
  "google-auth>=2.40.1,<3.0",
  "python-dotenv>=1.1.0,<2.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["fetch", "processing", "visualization"]
